Refactor TilemapGrid helpers
