+++ components/TilemapGrid.tsx
@@ -15,22 +15,15 @@ import {
   DAY_PHASE_CONFIG,
 } from "../lib/map";
 import type { Enemy } from "../lib/enemy";
-import type { NPC, NPCInteractionEvent } from "../lib/npc";
+import type { NPCInteractionEvent } from "../lib/npc";
 import { canSee, calculateDistance } from "../lib/line_of_sight";
-import { Tile } from "./Tile";
 import {
   getEnemyIcon,
   createEmptyByKind,
   EnemyRegistry,
-  type EnemyKind,
 } from "../lib/enemies/registry";
 import MobileControls from "./MobileControls";
 import styles from "./TilemapGrid.module.css";
-import {
-  computeTorchGlow,
-  ADJACENT_GLOW,
-  DIAGONAL_GLOW,
-} from "../lib/torch_glow";
 import { useRouter } from "next/navigation";
 // Daily flow is handled by parent via onDailyComplete when isDailyChallenge is true
 import { trackGameComplete, trackUse, trackPickup } from "../lib/analytics";
@@ -59,6 +52,8 @@ import { applyStoryEffectsWithDiary } from "../lib/story/event_registry";
 import { updateConditionalNpcs } from "../lib/story/story_mode";
 import { HeroDiaryModal } from "./HeroDiaryModal";
 import DayNightMeter from "./DayNightMeter";
+import { cloneDialogueLines } from "../lib/story/cloneDialogueLines";
+import { calculateMapTransform, renderTileGrid } from "./tilemapGridHelpers";
 
 type DialogueSession = {
   event: NPCInteractionEvent;
@@ -67,22 +62,6 @@ type DialogueSession = {
   dialogueId: string;
   consumedScriptIds: string[];
 };
-
-const TORCH_CARRIER_ENEMIES = new Set<EnemyKind>(["goblin"]);
-
-function cloneDialogueLines(lines: DialogueLine[]): DialogueLine[] {
-  return lines.map((line) => ({
-    ...line,
-    options: line.options
-      ? line.options.map((option) => ({
-          ...option,
-          response: option.response
-            ? cloneDialogueLines(option.response)
-            : undefined,
-        }))
-      : undefined,
-  }));
-}
 
 // Grid dimensions will be derived from provided map data
 
@@ -2433,417 +2412,3 @@ export const TilemapGrid: React.FC<TilemapGridProps> = ({
   );
 };
 
-// Calculate visibility based on player position
-function calculateVisibility(
-  grid: number[][],
-  playerPosition: [number, number] | null,
-  showFullMap: boolean = false,
-  heroTorchLit: boolean = true
-): number[][] {
-  const gridHeight = grid.length;
-  const gridWidth = grid[0].length;
-
-  // If showFullMap is true or no player, everything is visible
-  if (showFullMap || !playerPosition) {
-    return Array(gridHeight)
-      .fill(0)
-      .map(() => Array(gridWidth).fill(3));
-  }
-
-  // Create a grid of false values (not visible)
-  const visibility: number[][] = Array(gridHeight)
-    .fill(0)
-    .map(() => Array(gridWidth).fill(0));
-
-  const [playerY, playerX] = playerPosition;
-  // If hero torch is out, only reveal the hero's own tile; rely on wall torches for the rest
-  if (!heroTorchLit) {
-    // Center tile fully visible
-    visibility[playerY][playerX] = 3;
-    // Adjacent (orthogonal and diagonal) tiles dimly visible (tier 1)
-    const neighbors: Array<[number, number]> = [
-      // orthogonal
-      [playerY - 1, playerX],
-      [playerY + 1, playerX],
-      [playerY, playerX - 1],
-      [playerY, playerX + 1],
-      // diagonals
-      [playerY - 1, playerX - 1],
-      [playerY - 1, playerX + 1],
-      [playerY + 1, playerX - 1],
-      [playerY + 1, playerX + 1],
-    ];
-    for (const [y, x] of neighbors) {
-      if (y >= 0 && y < gridHeight && x >= 0 && x < gridWidth) {
-        visibility[y][x] = Math.max(visibility[y][x], 1);
-      }
-    }
-    return visibility;
-  }
-
-  // Full-visibility radius when torch is lit
-  const fullRadius = 4;
-
-  // Set visibility for all tiles in range
-  for (let y = 0; y < gridHeight; y++) {
-    for (let x = 0; x < gridWidth; x++) {
-      // Use Euclidean distance for circular FOV
-      const dy = y - playerY;
-      const dx = x - playerX;
-      const d = Math.sqrt(dx * dx + dy * dy);
-
-      // Tiered visibility:
-      // 3: d <= 4 (full)
-      // 2: 4 < d <= 5 (mid)
-      // 1: 5 < d <= 6 (low)
-      // 0: d > 6 (invisible)
-      let tier = 0;
-      if (d <= fullRadius) tier = 3;
-      else if (d <= fullRadius + 1) tier = 2;
-      else if (d <= fullRadius + 2) tier = 1;
-      else tier = 0;
-
-      visibility[y][x] = tier;
-    }
-  }
-
-  return visibility;
-}
-
-// Render the grid of tiles
-function renderTileGrid(
-  grid: number[][],
-  tileTypes: Record<number, TileType>,
-  subtypes: number[][][] | undefined,
-  environment: EnvironmentId | undefined,
-  showFullMap: boolean = false,
-  playerDirection: Direction = Direction.DOWN,
-  enemies?: Enemy[],
-  npcs?: NPC[],
-  hasSword?: boolean,
-  hasShield?: boolean,
-  heroTorchLit: boolean = true,
-  suppressDarknessOverlay: boolean = false,
-  hasExitKey?: boolean,
-  heroPoisoned: boolean = false,
-  timeOfDay?: TimeOfDayState,
-  activeCheckpoint?: [number, number] | null
-) {
-  const resolvedEnvironment = environment ?? DEFAULT_ENVIRONMENT;
-  const resolvedTimeOfDay = timeOfDay ?? createInitialTimeOfDay();
-  const timeOfDayVisual = DAY_PHASE_CONFIG[resolvedTimeOfDay.phase];
-  // Find player position in the grid
-  let playerPosition: [number, number] | null = null;
-
-  if (subtypes) {
-    for (let y = 0; y < subtypes.length; y++) {
-      for (let x = 0; x < subtypes[y].length; x++) {
-        if (subtypes[y][x].includes(TileSubtype.PLAYER)) {
-          playerPosition = [y, x];
-          break;
-        }
-      }
-      if (playerPosition) break;
-    }
-  }
-
-  // Calculate visibility for each tile, honoring hero torch state
-  const heroTorchLitForVisibility = suppressDarknessOverlay ? true : heroTorchLit;
-
-  const visibility = calculateVisibility(
-    grid,
-    playerPosition,
-    showFullMap,
-    heroTorchLitForVisibility
-  );
-
-  // Precompute torch glow positions by scanning for WALL_TORCH subtypes
-  const glowMap = new Map<string, number>();
-  const torchCarrierPositions = new Set<string>();
-  if (!suppressDarknessOverlay) {
-    if (subtypes) {
-      for (let y = 0; y < subtypes.length; y++) {
-        for (let x = 0; x < subtypes[y].length; x++) {
-          const st = subtypes[y][x];
-          if (st && st.includes(TileSubtype.WALL_TORCH)) {
-            const m = computeTorchGlow(y, x, grid);
-            for (const [k, v] of m.entries()) {
-              const prev = glowMap.get(k) ?? 0;
-              glowMap.set(k, Math.max(prev, v));
-            }
-          }
-        }
-      }
-    }
-
-    if (enemies) {
-      for (const enemy of enemies) {
-        if (!TORCH_CARRIER_ENEMIES.has(enemy.kind as EnemyKind)) continue;
-        torchCarrierPositions.add(`${enemy.y},${enemy.x}`);
-        const m = computeTorchGlow(enemy.y, enemy.x, grid);
-        for (const [k, v] of m.entries()) {
-          const prev = glowMap.get(k) ?? 0;
-          glowMap.set(k, Math.max(prev, v));
-        }
-      }
-    }
-  }
-
-  // Function to safely get a tile ID at specific coordinates
-  const getTileAt = (row: number, col: number): number | null => {
-    if (row < 0 || row >= grid.length || col < 0 || col >= grid[0].length) {
-      return null; // Out of bounds
-    }
-    return grid[row][col];
-  };
-
-  // Map enemies by position for sprite/facing lookup
-  const enemyMap = new Map<string, Enemy>();
-  if (enemies) {
-    for (const e of enemies) enemyMap.set(`${e.y},${e.x}`, e);
-  }
-
-  const npcMap = new Map<string, NPC>();
-  if (npcs) {
-    for (const npc of npcs) {
-      npcMap.set(`${npc.y},${npc.x}`, npc);
-    }
-  }
-
-  // Poison status passed in by caller
-
-  const tiles = grid.flatMap((row, rowIndex) =>
-    row.map((tileId, colIndex) => {
-      const tileType = tileTypes[tileId];
-      const subtype =
-        subtypes && subtypes[rowIndex] ? subtypes[rowIndex][colIndex] : [];
-      let tier = visibility[rowIndex][colIndex];
-      // Torch-driven illumination: use smaller radius similar to FOV tiers
-      const glowKey = `${rowIndex},${colIndex}`;
-      const g = glowMap.get(glowKey);
-      const isSelfTorch =
-        Array.isArray(subtype) && subtype.includes(TileSubtype.WALL_TORCH);
-      const isTorchCarrier = torchCarrierPositions.has(`${rowIndex},${colIndex}`);
-      if (isSelfTorch || isTorchCarrier) tier = Math.max(tier, 3);
-      // Neighbor illumination based on glow strength
-      if (g === ADJACENT_GLOW) {
-        tier = Math.max(tier, 2);
-      } else if (g === DIAGONAL_GLOW) {
-        tier = Math.max(tier, 1);
-      }
-      const isVisible = tier > 0;
-
-      // Get neighboring tiles
-      const neighbors = {
-        top: getTileAt(rowIndex - 1, colIndex),
-        right: getTileAt(rowIndex, colIndex + 1),
-        bottom: getTileAt(rowIndex + 1, colIndex),
-        left: getTileAt(rowIndex, colIndex - 1),
-      };
-
-      // Check if this is the player tile to pass the playerDirection prop
-      const isPlayerTile = subtype && subtype.includes(TileSubtype.PLAYER);
-
-      const enemyAtTile = enemyMap.get(`${rowIndex},${colIndex}`);
-      const hasEnemy = !!enemyAtTile;
-      const npcAtTile = npcMap.get(`${rowIndex},${colIndex}`);
-      const npcInteractable = (() => {
-        if (!npcAtTile || !playerPosition) return false;
-        if (npcAtTile.isDead()) return false;
-        const [py, px] = playerPosition;
-        return Math.abs(npcAtTile.y - py) + Math.abs(npcAtTile.x - px) === 1;
-      })();
-
-      return (
-        <div
-          key={`${rowIndex}-${colIndex}`}
-          className={`relative ${styles.tileWrapper}`}
-          style={(() => {
-            // Raise z-index so lit tiles appear above the dark vignette overlay
-            if (g != null || isSelfTorch) {
-              return { zIndex: 10050 as number };
-            }
-            return undefined;
-          })()}
-          data-row={rowIndex}
-          data-col={colIndex}
-        >
-          <Tile
-            tileId={tileId}
-            tileType={tileType}
-            subtype={subtype}
-            row={rowIndex}
-            col={colIndex}
-            isVisible={isVisible}
-            visibilityTier={tier}
-            neighbors={neighbors}
-            playerDirection={isPlayerTile ? playerDirection : undefined}
-            heroTorchLit={heroTorchLit}
-            heroPoisoned={isPlayerTile ? heroPoisoned : false}
-            hasEnemy={hasEnemy}
-            enemyVisible={isVisible}
-            enemyFacing={enemyAtTile?.facing}
-            enemyKind={
-              enemyAtTile?.kind as
-                | "goblin"
-                | "ghost"
-                | "stone-exciter"
-                | "snake"
-                | undefined
-            }
-            enemyMoved={Boolean(
-              (enemyAtTile?.behaviorMemory as Record<string, unknown> | undefined)?.["moved"]
-            )}
-            enemyAura={(() => {
-              if (!enemyAtTile) return false;
-              if (enemyAtTile.kind !== "stone-exciter") return false;
-              if (!playerPosition) return false;
-              const d =
-                Math.abs(enemyAtTile.y - playerPosition[0]) +
-                Math.abs(enemyAtTile.x - playerPosition[1]);
-              return d <= 2;
-            })()}
-            npc={npcAtTile}
-            npcVisible={npcAtTile ? isVisible : undefined}
-            npcInteractable={npcInteractable}
-            hasSword={hasSword}
-            hasShield={hasShield}
-            invisibleClassName={
-              process.env.NODE_ENV === "test"
-                ? "bg-gray-900"
-                : !heroTorchLitForVisibility
-                ? "bg-black"
-                : undefined
-            }
-            playerHasExitKey={hasExitKey}
-            environment={resolvedEnvironment}
-            suppressDarknessOverlay={suppressDarknessOverlay}
-            activeCheckpoint={activeCheckpoint}
-          />
-        </div>
-      );
-    })
-  );
-
-  // Add a smooth radial gradient overlay centered on the player for continuous fade
-  if (playerPosition && !showFullMap) {
-    const [py, px] = playerPosition; // grid coords
-    const tileSize = 40; // px (w-10/h-10)
-    const centerX = (px + 0.5) * tileSize;
-    const centerY = (py + 0.5) * tileSize;
-    // Stronger, darker fade pulled slightly inward
-    const r0 = 3.8 * tileSize; // inner safe
-    const r1 = 4.4 * tileSize; // begin fade
-    const r2 = 5.0 * tileSize; // mid fade
-    const r3 = 5.6 * tileSize; // stronger fade
-    const r4 = 6.2 * tileSize; // near dark
-    const r5 = 7.0 * tileSize; // full dark
-
-    // Warm torch glow radii (expanded for more dramatic effect)
-    const t0 = 2.5 * tileSize; // bright core (larger)
-    const t1 = 3.8 * tileSize; // warm mid (expanded)
-    const t2 = 5.2 * tileSize; // outer falloff (expanded)
-    const t3 = 6.5 * tileSize; // outer glow
-    const t4 = 7.5 * tileSize; // transparent edge (much larger)
-
-    const torchGradient = `radial-gradient(circle at ${centerX}px ${centerY}px,
-      var(--torch-core) ${t0}px,
-      var(--torch-mid) ${t1}px,
-      var(--torch-falloff) ${t2}px,
-      var(--torch-outer) ${t3}px,
-      rgba(0,0,0,0) ${t4}px
-    )`;
-
-    // When the hero's torch is OFF, use a pure black vignette to avoid gray tint.
-    // Otherwise, keep the dark gray for a softer ambiance.
-    const gradient = heroTorchLitForVisibility
-      ? `radial-gradient(circle at ${centerX}px ${centerY}px,
-      rgba(26,26,26,0) ${r0}px,
-      rgba(26,26,26,0.25) ${r1}px,
-      rgba(26,26,26,0.50) ${r2}px,
-      rgba(26,26,26,0.75) ${r3}px,
-      rgba(26,26,26,0.90) ${r4}px,
-      rgba(26,26,26,1) ${r5}px
-    )`
-      : `radial-gradient(circle at ${centerX}px ${centerY}px,
-      rgba(0,0,0,0) ${r0}px,
-      rgba(0,0,0,0.25) ${r1}px,
-      rgba(0,0,0,0.50) ${r2}px,
-      rgba(0,0,0,0.75) ${r3}px,
-      rgba(0,0,0,0.90) ${r4}px,
-      rgba(0,0,0,1) ${r5}px
-    )`;
-
-    // Push the warm torch glow first (lower z) ONLY if the hero's torch is lit,
-    // then the dark vignette (higher z) ONLY when torch is lit as well.
-    if (!suppressDarknessOverlay && heroTorchLitForVisibility) {
-      tiles.push(
-        <div
-          key="torch-glow"
-          className={`${styles.torchGlow}`}
-          style={{ backgroundImage: torchGradient, zIndex: 9000 }}
-        />
-      );
-    }
-    if (!suppressDarknessOverlay && heroTorchLitForVisibility) {
-      tiles.push(
-        <div
-          key="fov-radial-overlay"
-          className="pointer-events-none absolute inset-0"
-          style={{ backgroundImage: gradient, zIndex: 10000 }}
-        />
-      );
-    }
-    // Add a subtle green vignette when poisoned to suggest corrupted vision
-    if (heroPoisoned) {
-      tiles.push(
-        <div
-          key="poison-vignette"
-          className="poison-vignette"
-          style={{ zIndex: 10020 }}
-        />
-      );
-    }
-  }
-
-  if (timeOfDayVisual?.overlay) {
-    const overlayStyle: React.CSSProperties = {
-      background: timeOfDayVisual.overlay.color,
-      opacity: timeOfDayVisual.overlay.opacity,
-      zIndex: 8800,
-    };
-    if (timeOfDayVisual.overlay.blendMode) {
-      overlayStyle.mixBlendMode = timeOfDayVisual.overlay.blendMode;
-    }
-    tiles.push(
-      <div
-        key="time-of-day-overlay"
-        className="pointer-events-none absolute inset-0"
-        style={overlayStyle}
-      />
-    );
-  }
-
-  return tiles;
-}
-
-// Function removed since we're now using GameState
-
-// Calculate the transform to center the map on the player
-function calculateMapTransform(playerPosition: [number, number]): string {
-  if (!playerPosition) return "0px, 0px";
-
-  const tileSize = 40; // px
-  const viewportWidth = 600; // px (from CSS)
-  const viewportHeight = 600; // px (from CSS)
-
-  // Calculate the center position of the player in pixels
-  const playerX = (playerPosition[1] + 0.5) * tileSize;
-  const playerY = (playerPosition[0] + 0.5) * tileSize;
-
-  // Calculate the transform to center the player in the viewport
-  const translateX = viewportWidth / 2 - playerX;
-  const translateY = viewportHeight / 2 - playerY;
-
-  return `${translateX}px, ${translateY}px`;
-}
